Better `import` and `export`
Made the communication interface better: no longer tries to interact with the global object, instead uses its own property, `ISLInterpreter.communication`, and no longer uses new Function().

Types can have `=` prefixing them to use exact words e.g. `=as|to` means `as` or `to`

`import` and `export` require `as` or `to` to specify add mode: `as` creates but throws if present, `to` is the opposite.

`getkeys`, `awaitkey` don't require a manipulator.

`getkeys`, `popup-input` and `awaitkey` require the word 'as' before the variable name: this is entirely to improve readability